Changed the sample size on data-source so you can change it in the menus. Also fixed an issue with CSV files where they did not use the sampling loading bar.
