+++ plugins/mongo/server/EBMongoPlugin.js
@@ -359,53 +359,6 @@ class EBMongoPlugin extends EBDataSourcePlugin
         });
     }
 
-
-    /**
-     * This method should sample the data from the data-source and determine what the schema is.
-     *
-     * @param {EBDataSource} dataSource This should be the EBDataSource object to find the tables with it
-     * @param {function(dataSchema, objectsCompleted, objectsTotal)} [iterator] This is an iterator function that is called with the intermediary schemas as they are being assembled.
-     * @returns {Promise} A promise that will resolve to the EBSchema object for this data-source.
-     */
-    detectSchema(dataSource, iterator)
-    {
-        const maxNumberOfObjectsToSample = 500;
-        const examplesToKeep = 10;
-        const objectsBetweenIteratorCalls = 25;
-        const exampleIndexes = randomUtilities.getRandomIntegers(maxNumberOfObjectsToSample, examplesToKeep);
-
-        const schemaDetector = new EBSchemaDetector(this.application);
-        let objectIndex = 0;
-
-        return this.sample(maxNumberOfObjectsToSample, dataSource,
-            (object) =>
-            {
-                let keepForSample = false;
-                if (exampleIndexes.indexOf(objectIndex) !== -1)
-                {
-                    keepForSample = true;
-                }
-
-                // Accumulate the converted version of the object
-                objectIndex += 1;
-
-                let iteratorPromise = Promise.resolve();
-                if (objectIndex % objectsBetweenIteratorCalls === 0)
-                {
-                    iteratorPromise = iterator(schemaDetector.getSchema(), objectIndex, maxNumberOfObjectsToSample);
-                }
-
-                return iteratorPromise.then(() =>
-                {
-                    return schemaDetector.accumulateObject(object, keepForSample);
-                });
-            }).then(() =>
-            {
-                const schema = schemaDetector.getSchema();
-                return schema;
-            });
-    }
-
     /**
      * This static function is used to convert objects that come out of Mongo into pure JSON
      *
